+++ src/Fp.Tests/ApiTests.cs
@@ -113,7 +113,7 @@ public class ApiTests
     {
         byte[] data;
         byte[] tmp = File.ReadAllBytes("Watch_Dogs2020-4-3-0-57-53.png");
-        data = new byte[Processor.GetPaddedLength(tmp.Length, Processor.PaddingMode.Zero, 8)];
+        data = new byte[Processor.GetPaddedLength(tmp.Length, CipherPaddingMode.Zero, 8)];
         tmp.AsSpan().CopyTo(data);
 
         byte[] dataEnc = new byte[data.Length];